requirements.txt
langchain-ollama==0.2.2
fastapi==0.115.6
pydantic==2.10.5
uvicorn==0.34.0
ipython==8.31.0
pillow==11.1.0
langchain-core==0.3.31
transformers==4.48.1
torch==2.5.1
torchvision==0.20.1